fix the jenga&VSA code review

+++ include/ck_tile/core/arch/amd_buffer_addressing_builtins.hpp
@@ -44,9 +44,9 @@ __device__ inline int32_t amd_wave_read_first_lane(int32_t value)
     return __builtin_amdgcn_readfirstlane(value);
 }
 
-__device__ inline int32_t amd_wave_read_first_lane(uintptr_t value)
+__device__ inline uint32_t amd_wave_read_first_lane(uintptr_t value)
 {
-    return __builtin_amdgcn_readfirstlane(value);
+    return __builtin_amdgcn_readfirstlane(static_cast<uint32_t>(value));
 }
 
 template <typename Object, std::enable_if_t<std::is_trivially_copyable_v<Object>, int> = 0>
